Replace climate model normals with 20-year historical archive and show tenths for all days
Climate normals are now derived from a rolling 20-year window of observed
daily highs/lows from the Open-Meteo historical archive (ERA5 reanalysis),
replacing the single-year climate-model projection that ran several degrees
hot. Monthly means are cached as 12 rows and expanded to per-day values via
linear interpolation between month midpoints.

Database migration 46→47 drops and recreates climate_normals with REAL
columns (Float) instead of INTEGER, wiping stale values so they refetch.

Temperature labels now show the tenth for all days (today, past, and future)
with ".0" suppression for whole-degree sources. formatTempLabel loses the
isActualData flag in favor of a forceInteger parameter for low-label
collision avoidance with night-rain labels.

OpenMeteoApi.getClimateForecast → getHistoricalDailyTemps, switching from
the climate-projection endpoint to the archive endpoint with auto timezone.

+++ app/src/main/java/com/weatherwidget/data/local/ClimateNormalEntity.kt
@@ -8,12 +8,16 @@ import androidx.room.Query
 
 @Entity(tableName = "climate_normals", primaryKeys = ["monthDay", "locationKey"])
 data class ClimateNormalEntity(
-    /** "MM-dd" format, e.g. "01-15" for January 15. */
+    /**
+     * "MM-dd" format. Stored as a mid-month anchor ("MM-15", e.g. "06-15"): one row
+     * per month holds that month's mean, expanded to per-day values by interpolation
+     * when read.
+     */
     val monthDay: String,
     /** Rounded lat/lon key, e.g. "37.4_-122.1". */
     val locationKey: String,
-    val highTemp: Int,
-    val lowTemp: Int,
+    val highTemp: Float,
+    val lowTemp: Float,
     val fetchedAt: Long = System.currentTimeMillis(),
 )
